requirements.txt
Django==2.1
psycopg2
psycopg2-binary
djangorestframework
markdown
django-filter
requests
python-dateutil